requirements.txt
langchain
langchain-groq
langchain_community
sqlalchemy
pandas
numpy
groq
python-dotenv
ipykernel
psycopg2
streamlit